pyproject.toml
[project]
name = "vex-ads"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.109.0",
    "google-genai>=1.0.0",
    "httpx>=0.26.0",
    "pillow>=10.2.0",
    "pydantic-settings>=2.1.0",
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "python-dotenv>=1.0.0",
    "python-multipart>=0.0.6",
    "supabase>=2.0.0",
    "uvicorn[standard]>=0.27.0",
]
